Move C++ `GlobalType` to its own header
